refactor(event): extract shared unix parser into its own module
`src/event/source/unix/tty.rs` and `src/event/source/unix/mio.rs` each carried their own copy of a `Parser` struct that buffers bytes across `read()` boundaries and dispatches them to `parse_event`. The two copies were identical, including comments.

Move the canonical implementation into a new `src/event/source/unix/parser` module and have both event sources import `Parser` from there.

No effect on behaviour.
